Izlaboju Apskatit logu un Main8
