+++ cenas.txt
@@ -1,5 +1,2 @@
 4.5
-0.3
-0.25
-0.55
-5.0
+3.0
