+++ izveletieToppingi.txt
@@ -1,3 +1,2 @@
-Siers
 Sēnes
 Maltā gaļa
