+++ pagaidu_pasutijuma_dati.txt
@@ -1,0 +1,10 @@
+Nosaukums: wadwad, Veids: Lauku, Toppings: [Sēnes, Maltā gaļa], Picas Diametrs: 50cm 
+Piegāde informācija:
+Vārds: dawd, Uzvārds: dawdaw, Tel. nr.: adawd, Adrese: adawd, Pasta Indekss: LV-dawd
+Kopēja cena 10.3€
+----------------------------------------------------------------------------------
+Nosaukums: dddddd, Veids: Lauku, Toppings: [Sēnes, Maltā gaļa], Picas Diametrs: 30cm 
+Piegāde informācija:
+Vārds: dawda, Uzvārds: dadaw, Tel. nr.: awdaw, Adrese: dawdaw, Pasta Indekss: LV-daww
+Kopēja cena 7.5€
+----------------------------------------------------------------------------------
