+++ pastaIndekss.txt
@@ -1,1 +1,1 @@
-awda
+daww
