--- pasutijuma_dati.txt
@@ -1,6 +1,0 @@
-Pica: dawdawd, Veids: Lauku, Toppings: [Siers, Sēnes, Maltā gaļa], Picas Diametrs: 50
- Piegāde informācija:
-Vārds: awdad, Uzvārds: awdawda, Tel. nr.: adawd, Adrese: awdawd, Pasta Indekss: adad
-Pica: dawdaw, Veids: Lauku, Toppings: [Siers, Sēnes], Picas Diametrs: 50
- Piegāde informācija:
-Vārds: dada, Uzvārds: dadawdaw, Tel. nr.: dawdawd, Adrese: dawdawd, Pasta Indekss: dawd
