+++ picasDiametrs.txt
@@ -1,1 +1,1 @@
-50
+30
