+++ src/lvt/Apskatit.java
@@ -12,7 +12,6 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
-
 import javax.swing.DefaultListModel;
 import javax.swing.ImageIcon;
 import javax.swing.JButton;
@@ -20,7 +19,6 @@ import javax.swing.JFrame;
 import javax.swing.JList;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
-import javax.swing.JScrollBar;
 import javax.swing.JScrollPane;
 import javax.swing.border.EmptyBorder;
 import javax.swing.border.LineBorder;
@@ -59,8 +57,7 @@ public class Apskatit extends JFrame {
 
  public Apskatit() {
 	 
-	 pica picasObjekts = nolasitUnSaglabatInformaciju();
-	 ierakstitObjektuFaila(picasObjekts);
+	 
     	
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         setBounds(100, 100, 800, 800);
@@ -139,112 +136,7 @@ public class Apskatit extends JFrame {
     }
 
     
-    public pica nolasitUnSaglabatInformaciju() {
-        String pasutijumaNosaukums = "";
-        String picasVeids = "";
-        ArrayList<String> izveletieToppingi = new ArrayList<>();
-        String picasDiametrs = "";
-        String vards = "";
-        String uzvards = "";
-        String telNr = "";
-        String adrese = "";
-        String pastaIndekss = "";
-
-        try (BufferedReader br = new BufferedReader(new FileReader("pasutijumaNosaukums.txt"))) {
-            pasutijumaNosaukums = br.readLine();
-        } catch (IOException e) {
-            e.printStackTrace();
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu pasutijumaNosaukums.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-
-        try (BufferedReader br = new BufferedReader(new FileReader("picasVeids.txt"))) {
-            picasVeids = br.readLine();
-        } catch (IOException e) {
-            e.printStackTrace();
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu picasVeids.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-
-        try (BufferedReader br = new BufferedReader(new FileReader("izveletieToppingi.txt"))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] toppings = line.split(","); 
-                for (String topping : toppings) {
-                    izveletieToppingi.add(topping.trim());
-                }
-            }
-        } catch (IOException e) {
-         
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu izveletieToppingi.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-
-        try (BufferedReader br = new BufferedReader(new FileReader("picasDiametrs.txt"))) {
-            picasDiametrs = br.readLine();
-        } catch (IOException e) {
-            
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu picasDiametrs.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-
-        try (BufferedReader br = new BufferedReader(new FileReader("vards.txt"))) {
-            vards = br.readLine();
-        } catch (IOException e) {
-           
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu vards.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-
-        try (BufferedReader br = new BufferedReader(new FileReader("uzvards.txt"))) {
-            uzvards = br.readLine();
-        } catch (IOException e) {
-          
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu uzvards.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-        
-        try (BufferedReader br = new BufferedReader(new FileReader("telNr.txt"))) {
-            telNr = br.readLine();
-        } catch (IOException e) {
-           
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu uzvards.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-        
-        try (BufferedReader br = new BufferedReader(new FileReader("adrese.txt"))) {
-            adrese = br.readLine();
-        } catch (IOException e) {
-         
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu uzvards.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-        
-        try (BufferedReader br = new BufferedReader(new FileReader("pastaIndekss.txt"))) {
-            pastaIndekss = br.readLine();
-        } catch (IOException e) {
-          
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu uzvards.txt", "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-        double summa = 0.0;
-        try (BufferedReader br = new BufferedReader(new FileReader(CENAS_FAILS))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                double cena = Double.parseDouble(line.trim());
-                summa += cena;
-            }
-        } catch (IOException e) {
-            JOptionPane.showMessageDialog(null, "Kļūda nolasot failu " + CENAS_FAILS, "Kļūda", JOptionPane.ERROR_MESSAGE);
-        }
-        
-     
-        
-      
-        return new pica(pasutijumaNosaukums, picasVeids, izveletieToppingi, picasDiametrs, vards, uzvards, telNr, adrese, pastaIndekss, summa);
-    }
-  
- 
-
-        private static void ierakstitObjektuFaila(pica pica) {
-            try (BufferedWriter wr = new BufferedWriter(new FileWriter(FAILS, true))) {
-                wr.write(pica.toString());
-                wr.newLine();
-            } catch (IOException e) {
-                JOptionPane.showMessageDialog(null, "Kļūda ierakstot objektu failā " + FAILS, "Kļūda", JOptionPane.ERROR_MESSAGE);
-            }
-        }
+   
         private void izdzestPasutijumus() {
             try (BufferedWriter wr = new BufferedWriter(new FileWriter(FAILS, false))) {
                 wr.write("");
