+++ telNr.txt
@@ -1,1 +1,1 @@
-dawdawd
+awdaw
