+++ uzvards.txt
@@ -1,1 +1,1 @@
-dawdaw
+dadaw
